[#17] supporters.py: houses the classes that support the main commands. Downloader &  Splitter have been defined. Still need to work on Indexer &  Searcher
