+++ requirements.txt
@@ -1,0 +1,53 @@
+beautifulsoup4==4.10.0
+cachetools==4.2.4
+certifi==2021.10.8
+charset-normalizer==2.0.7
+click==8.0.3
+configparser==5.0.2
+docker-pycreds==0.4.0
+elasticsearch==7.15.0
+gitdb==4.0.7
+GitPython==3.1.24
+google-api-core==2.1.1
+google-auth==2.3.0
+google-cloud-core==2.1.0
+google-cloud-storage==1.42.3
+google-crc32c==1.3.0
+google-resumable-media==2.0.3
+googleapis-common-protos==1.53.0
+httplib2==0.20.1
+idna==3.3
+importlib-metadata==4.8.1
+joblib==1.1.0
+numpy==1.21.2
+oauth2client==4.1.3
+pandas==1.3.3
+pathtools==0.1.2
+promise==2.3
+protobuf==3.18.1
+psutil==5.8.0
+pyasn1==0.4.8
+pyasn1-modules==0.2.8
+pyparsing==2.4.7
+python-dateutil==2.8.2
+python-dotenv==0.19.1
+pytz==2021.3
+PyYAML==6.0
+requests==2.26.0
+rsa==4.7.2
+scikit-learn==1.0
+scipy==1.7.1
+sentry-sdk==1.4.3
+shortuuid==1.0.1
+six==1.16.0
+smmap==4.0.0
+soupsieve==2.2.1
+subprocess32==3.5.4
+termcolor==1.1.0
+threadpoolctl==3.0.0
+tqdm==4.62.3
+typing-extensions==3.10.0.2
+urllib3==1.26.7
+wandb==0.12.4
+yaspin==2.1.0
+zipp==3.6.0
